requirements.txt
python-telegram-bot>=20.0
supabase>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
faster-whisper>=0.9.0
pydub>=0.25.1
